expiration time warnings

+++ app/controllers/spree/base_controller_decorator.rb
@@ -2,6 +2,8 @@ Spree::BaseController.class_eval do
   prepend_before_filter :seek_and_destroy_expired_stock_keepings
 
   def seek_and_destroy_expired_stock_keepings
-    current_order.destroy if current_order(true).stock_keepings_expired?
+    if self.respond_to? :current_order
+      current_order.destroy if current_order(true).stock_keepings_expired?
+    end
   end
 end
